LAB-1_all_Program_27-05-2021
I have written all the programs in a single file which covered in first class on date 27-05-2021.

Q. 1 Create two integer variables and do basic arithmetic operations and print the values. 
Q.2 Create a variable called DNA and store "ATACGATACAA", print dn.
Q. 3 Create a class with the list of variables with names of amino acids and store the molecular weights and print amino acid and its weight. 
Q. 4 Create a string array with 20 amino acids and print them in alphabetical order. 
Q. 5 Create a double array of 10 elements & find its average. 
Q. 6 Create a class with 5 subjects and calculate the % and if the % is above 50 print pass. 
Q. 7 Print fibonacci series.
Q. 8 Create a double array of 10 elements and find its average.